+++ receiver/hostmetricsreceiver/internal/scraper/loadscraper/load_scraper_test.go
@@ -66,7 +66,7 @@ func TestScrape(t *testing.T) {
 
 	for _, test := range testCases {
 		t.Run(test.name, func(t *testing.T) {
-			scraper := newLoadScraper(context.Background(), zap.NewNop(), &Config{PerCPU: test.perCPU})
+			scraper := newLoadScraper(context.Background(), zap.NewNop(), &Config{CPUAverage: test.perCPU})
 			if test.loadFunc != nil {
 				scraper.load = test.loadFunc
 			}