+++ CHANGELOG.md
@@ -31,7 +31,7 @@
 - `tanzuobservabilityexporter`: Document how to enable memory_limiter (#7286)
 - `hostreceiver/networkscraper`: Migrate the scraper to the mdatagen metrics builder (#7048)
 - `hostmetricsreceiver`: Add MuteProcessNameError config flag to mute specific error reading process executable (#7176)
-- `hostmetricsreceiver`: add option for load scraper to average the value per cpus (#6999)
+- `hostmetricsreceiver`: add `cpu_average` option for load scraper to report the average cpu load (#6999)
 
 ## 🛑 Breaking changes 🛑